+++ frontend/src/App.scss
@@ -392,6 +392,16 @@ nav {
   align-items: center;
   margin: auto;
 }
+
+.ProgramSection {
+  .black-button {
+    margin: 32px auto;
+  }
+}
+
+.programdesc {
+  margin: 30px 0;
+}
 /* ---------------------------------------------------MANNEQUIN SECTION HOMEPAGE----------------------------------------------------------------- */
 
 .DesktopSwiper {
