--- frontend/src/components/Week.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-function Week() {
-  return (
-    <>
-      <h1>Week</h1>
-      <h2>hello</h2>
-    </>
-  );
-}
-
-export default Week;
